Add Qt6 qmake to candidates for Qt documentation lookup
On my system, and in the official Qt6 tarballs, the Qt6 qmake binary
is called qmake6, not qmake-qt6, so use that first. Maybe there
are other systems using the previous suffix kind, i.e. -qt6, so also
take that into account.

We do not yet prefer this version, as we cannot yet parse it
correctly, see the expected failure in the test.