Also log failure when registering documentation from QtHelpProvider
This was so far unhandled and only the QtHelpQtDoc logged the
failure case.

To prevent log-spam we don't emit a warning when the namespace
was already registered. This might hint at a problem though:
If the file source changes, we should probably unregister the
namespace and reload it anew. Doing that will have to happen
in a later patch.