Fix QtHelpPlugin for Qt6: make engine writable
Qt6 introduced the readOnly concept which is set to true by default.
This completely breaks out documentation plugin since we assume
the old Qt5 behavior of writable engines. Add the explicit call
here and verify in the test that we actually manage to load some data
from the qt documentation.