Fix QtHelp description extraction for new Qt6 format
Add two new heuristics to extract the short description from
the HTML embedded in the Qt help files:

First, we try to rely on the comment markers to find the section
we are interested in. Failing that, we try to parse the new structure
found in the Qt6 documentation. Failing that we fallback to the old
code.

Example contents on my system for Qt6:
```
<!-- $$$fromLatin1[overload1]$$$fromLatin1constchar*qsizetype -->
<h3 class="fn" translate="no" id="fromLatin1">...</h3>
<p>...</p>
<p>...</p>
<p>...</p>
<!-- @@@fromLatin1 -->
```