Actually show the Setup Custom Include Paths dialog
Clicking the "Setup Custom Include Paths" Solution link in a popup over
a missing include Problem in a file that does not belong to a project,
has no visible effect since 002fbc583524bbb662f47fcdbcbd0395f007a738.
A NoProjectCustomIncludePaths dialog is created but newer shown, and
therefore is leaked. This is because the aforementioned commit removed
a QDialog::exec() call but failed to replace it with a QDialog::show()
call.