ProjectController: pass a safer context to QObject::connect()
Both the ProjectControllerPrivate and a Project* proj are used in a
lambda, but the sender of the signal cfgDlg is passed as the context to
QObject::connect(), even though cfgDlg is not used in the lambda and a
signal is never emitted when its sender is dead anyway. This mistake was
introduced in 80bb2205945f53f683c96a38715f185965944ae1.

ProjectControllerPrivate outlives all Project objects, because it is
destroyed along with Core, on which deleteLater() is called after
Core::cleanup(), and ProjectController::cleanup() indirectly calls
deleteLater() on all projects during Core::cleanup(). As far as I know,
within a common thread, an object on which deleteLater() is called
earlier, is destroyed earlier too. Therefore, pass the shorter-lived
proj as the context.

The aforementioned commit also made an assertion within the lambda
fruitless, because the enclosing function early-returns if proj is null.
Remove the assertion.

Also remove the unused lambda parameter.