Convert class NoProjectIncludePathsManager into namespace
The class NoProjectIncludePathsManager is stateless, because it contains
no data members. Make this clearer by converting it into a namespace.
Give its private member functions internal linkage.

The NoProjectIncludePathsManager object could cache parsed includes and
defines. But an entire fruitful (when a .kdev_include_paths file is
found) call to NoProjectIncludePathsManager::includesAndDefines() takes
less than a millisecond (about 0.3 ms on average) on my old computer.
For comparison, the ClangParseJob constructor, from which
includesAndDefines() is invoked three times, rarely takes less than 2 ms
on the same computer. So the cache, with its extra complexity and risks,
is unlikely to be implemented any time soon.