NoProjectIncludePathsManager: show defines in text edit
NoProjectIncludePathsManager::openConfigurationDialog() passes the list
of includes - includesAndDefines(path).first - to the
NoProjectCustomIncludePaths dialog. So any defines possibly present in
the user's .kdev_include_paths file are silently filtered out and not
present in the initial text edit presented to the user.

Let the user edit the entire unprocessed contents of .kdev_include_paths
by skipping the include and define parsing in openConfigurationDialog().
This is both more correct and more efficient. Pass the entire file
contents QString instead of the parsed includes QStringList to
NoProjectCustomIncludePaths::setCustomIncludePaths().

Write the contents of the user-edited text edit verbatim to
.kdev_include_paths instead of splitting into lines and trivially
formatting them with QTextStream. This makes the code both simpler and
more efficient. And preserves empty lines, with which the user may
separate text blocks for readability.
NoProjectIncludePathsManager::includesAndDefines() trims and ignores
empty lines while parsing .kdev_include_paths, so writing them to the
file is fine. Trim the contents of the text edit before writing to file
in order to remove the file in case of whitespace-only text.

Return false from writeIncludePaths() if writing to file fails in order
to print the warning
"Failed to save custom include paths in directory: %1".

Reuse QDir and QFileInfo objects to optimize.

Call path() instead of absolutePath() on an absolute QDir to optimize.