Don't consider NoProjectIncludePathsManager ProjectSpecific
3512f8906466d9602f6faf8f0a64f9dd8d930201 disabled CompilerSpecific
defines and includes for QmlJS. It also started to consider
NoProjectIncludePathsManager's defines and includes ProjectSpecific in
one defines() and one includes() overload. The other overload of each
unconditionally adds NoProjectIncludePathsManager's defines and
includes. The ProjectSpecific change is unrelated and unneeded for the
fix implemented by the aforementioned commit.

Revert the ProjectSpecific change to consistently unconditionally add
NoProjectIncludePathsManager's defines and includes in all overloads.
The ProjectSpecific flag is always passed to includes() and defines(),
therefore this change does not affect KDevelop's behavior.