debuggercommon: don't call QStringView::mid(-1, 2)
QStringView::mid() will be ported to QStringView::sliced() once KDevelop
depends on Qt 6. The behavior is undefined if any argument to sliced()
is negative.

index can be equal to 0 at the modified line. A recent commit ported
QString::midRef(index - 1, 2) to the current
QStringView::mid(index - 1, 2). midRef(-1, 2) returns the single first
character of the string. All usages of ch2 compare it to a string of
size 2. The result of such a comparison is the same whether ch2.size()
is 1 or 0. Assign an empty QStringView to ch2 instead of the first
character in this case to simplify code without affecting behavior.