GitPlugin: prepare to QStringView's first() and sliced()
QStringView::left() will be ported to QStringView::first() and
QStringView::mid() - to QStringView::sliced() once KDevelop depends on
Qt 6. The behavior is undefined if an argument to first() is negative
or greater than the string view's size. The behavior is also undefined
if any argument to sliced() is negative or the argument to the
single-parameter overload of sliced() is greater than the string view's
size.

Do not assert that the output of a git command has the expected format,
because the format can conceivably change in a future git version. Print
a warning and skip a line that has an unexpected format.