GitPlugin::parseGitBranchOutput: prepare to QStringView::sliced()
QStringView::mid() will be ported to QStringView::sliced() once KDevelop
depends on Qt 6. The behavior is undefined if the argument to sliced()
is greater than the string view's size.

The current branch is marked with an asterisk in the output of the
`git branch` command. The asterisk is followed with a space. The
QStringView branch is trimmed before its single remaining use just
below, so there is no need to remove the space after the asterisk as it
is going to be removed during the trimming.