PatchHighlighter: assert KTextEditor cursor assumptions
PatchHighlighter::textRemoved() and PatchHighlighter::textInserted() are
connected to namesake KTextEditor::Document signals. So the range and
cursor arguments must be valid and columns must be valid in the
corresponding lines.

QStringView::left() will be ported to QStringView::first() and
QStringView::mid() - to QStringView::sliced() once KDevelop depends on
Qt 6. The behavior is undefined if the argument to first() or sliced()
is negative or greater than the string view's size.

Assert the KTextEditor cursor assumptions to better document them and
catch potential bugs earlier.