GitPlugin: store encountered revisions in QHash rather than QMap
Neither the iterator or reference stability nor the order of elements in
this associative container matters. QHash is usually much faster than
QMap.

Convert a QStringView to QString once rather than 3 times.

Reduce the number of associative container lookups using the fact that
a default-constructed VcsRevision's type is Invalid.

Create and set VcsRevision once for each VcsAnnotationLine. The revision
type is constant (GlobalNumber) and the revision value is uniquely
determined by the looked-up key to definedRevisions (name).