GitPlugin: ignore the "boundary" revision marker
From git-blame documentation:
When revision range specifiers are used to limit the annotation, lines
that have not changed since the range boundary are blamed for that range
boundary commit.

KDevelop never limits annotation to a revision range, so the "boundary"
designation of a commit is uninteresting. Inserting {"boundary", {}}
into definedRevisions does not affect behavior, because this inserted
element is not referenced anywhere else. The commit that reimplemented
GitPlugin::parseGitBlameOutput() -
98d2b3d0419589b4301ce505a8f491561bec4f4a - introduced this insertion
without explanation.