Unbreak `QtHelpDocumentation::description` title matching
The recent porting of the variable `between` from QStringRef to
QStringView exposed an old bug by consistently crashing
test_qthelpplugin:
    QFATAL : TestQtHelpPlugin::testDeclarationLookup_Method() ASSERT:
    "len >= 0" in file /usr/include/qt/QtCore/qstringview.h, line 175

The same assertion also fails in KDevelop itself occasionally.

Qt developers' intention probably was to make QStringView::left() and
mid() fully compatible with QString[Ref]::left() and mid(), but in
practice compatibility only covers the documented behavior in Qt 5: the
only supported negative number is -1. The compatibility is full in Qt 6
thanks to code reuse.

Use a non-greedy `.*` capture in the `titleRegExp`, otherwise this
would match potentially much more than just a single header.

Then use `QString::lastIndexOf` to get the old reverse matching
behavior, which got accidentally removed when this was ported
from QRegularExpression to QRegExp in
65a3d93d6b8526b2a2d22e7c6daaa051b05734cc.

Finally fix the `titleEnd` initialization, `capturedEnd` is
the position directly, not the size that needs to be added
to the `titleStart`.