debugger: do not call KNotification::setWidget()
KNotification::setWidget() is called to make the Taskbar action work.
This action got removed in KF6 however, so there is no need to associate
a widget with a notification anymore.

KNotification::setWindow() replaces KNotification::setWidget() in KF6.
But the only use of setWindow() is to support CloseWhenWindowActivated
notification flag (according both to KNotification documentation and its
source code). KDevelop does not use or need this notification flag, so
just remove the setWidget() calls rather than port them to setWindow().