RegistersManager: don't cast pointer to bool before printing
The type of m_registerController is QScopedPointer<IRegisterController>.
QScopedPointer is not convertible to its raw pointer type, but is
implicitly convertible to bool in Qt version 5 and earlier. Such a
conversion to bool happens when a QScopedPointer is printed via QDebug
in Qt 5. This printing does not compile at all in Qt 6, because
QScopedPointer::operator bool() became explicit.

The conversion to bool before printing was accidentally introduced by
the commit that changed the type of m_registerController from raw
pointer to QScopedPointer: 1821513c2de8211d4cc3736af10af7c96fba85cd.

Improve formatting of the debug message while at it.