Port away from QAction::parentWidget()
QAction::parentWidget() is deprecated since Qt 6.0.

The implementation of QAction::parentWidget() does not merely cast
parent() to QWidget*. But, in case the parent is not a QWidget, casts a
grandparent, great-grandparent, and so on, to QWidget*, and returns
nullptr only if no ancestor of the QAction is a QWidget.

The slot QuickOpenWidget::updateProviders() is connected to only in
QuickOpenWidget(). Each action, whose signal is connected to the slot,
has a QMenu (which inherits QWidget) as the parent. qobject_cast to
QMenu* action->parent() instead of action->parentWidget(), because this
is correct and even more efficient here (fewer casts and checks).