QtHelpDocumentation: actually show function description
descriptionFromHtmlData() searches the nextFragmentExpression regex
starting from the index 0 in case of empty fragment (i.e. class
documentation), which is correct. In case of nonempty fragment (i.e.
function documentation), it should (and apparently aims to) search the
nextFragmentExpression regex starting at the end of the findFragment
regex match. But when a lower number titleStart is assigned to pos, the
search usually starts before the beginning of the findFragment regex
match, and the same fragment is found again. Only an invisible header
tag added via the findTitle regex remains between pos and endPos then.
Therefore, for most (maybe even all) Qt [member] functions, the
description displayed in a navigation widget is empty.

Store the correct end position of the findFragment regex match in a new
dedicated variable - nextFragmentSearchPos.

When the fragment is empty, descriptionFromHtmlData() searches for
the findFragment regex and then discards the search results. Prevent
this useless search in order to optimize.

Return earlier if the findFragment regex is not found (pos == -1). The
search for the findTitle regex just below does not make sense and
Q_ASSERT(titleStart < pos) would certainly fail in this case.

Change the type of pos from int to QString::size_type.
QString::size_type is int in Qt 5 and qsizetype in Qt 6.

Do not expect to find the member function name in its own description in
TestQtHelpPlugin::testDeclarationLookup_Method(). The test succeeded
before this commit, because the description consisted of the following
invisible header/caption opening tag:
"<big class=\"fn\" translate=\"no\" id=\"fromLatin1\">". It was not
matched by the findHeader regex and thus not removed, because the
closing tag was missing. Now that the caption is correctly matched and
removed, there is no "fromLatin1" string in the description anymore.
Expect to find a closely related member function name "toLatin1" on the
"See also" line instead in order to prevent the test failure.