QtHelpDocumentation: fix matching "title" class
translate="no" appears after class="title" in the documentations of
Qt 5.15.14 classes. This clutters the beginning of any Qt class
description in a navigation widget with the useless navigation bar and
table of contents. The class title such as "QString Class" is also shown
and duplicates the beginning of the navigation widget text such as
"class QString", because the table of contents title
"<h3><a name=\"toc\">Contents</a></h3>" is matched by the findHeader
regex and removed instead of e.g.
"<h1 class=\"title\" translate=\"no\">QString Class</h1>".