Don't call QRegularExpressionMatch::isValid()
The documentation for this function:
    Returns true if the match object was obtained as a result from the
    QRegularExpression::match() function invoked on a valid
    QRegularExpression object; returns false if the QRegularExpression
    was invalid.

The validity of a QRegularExpression is determined by its pattern string
and should not be checked at run-time if the pattern is known at
compile-time.

65a3d93d6b8526b2a2d22e7c6daaa051b05734cc and
b36004bd896f10a3673d97d09f0a845c0be49e1d introduced the usages without
explanation and probably by mistake - in place of
QRegularExpressionMatch::hasMatch().

Do not insert a useless empty target string in enumerateTargets() if
`line` does not match the regular expression.

Do not explicitly check the lastNewLine regex match success, because the
existing `newEnd > pos` check rejects the no-match condition
newEnd == -1 seeing as pos is nonnegative.

Do not explicitly check the findTitle regex match success, because the
call qBound(pos, titlePos, endPos) returns pos anyway in case of no
match (titlePos == -1; pos >= 0).

The added hasMatch() check for the findHeader regex does not affect
KDevelop behavior, because QString::remove(-1, 0) was called in case of
no match, and such a call does nothing in both Qt 5 and Qt 6.