Don't compare different specializations of QExplicitlySharedDataPointer
In Qt 5 QExplicitlySharedDataPointer's operator== and operator!= are
non-static data members. Therefore compilers do not consider implicit
conversions of the first argument to the comparison operators. So the
second argument to a comparison operator is converted from e.g.
IntegralType::Ptr to AbstractType::Ptr and the two data() pointers are
compared.

In Qt 6 QExplicitlySharedDataPointer's operator== and operator!= are
friend non-member functions. Therefore compilers consider implicit
conversions of both arguments to the comparison operators. This results
in an ambiguous overload compilation error:
/usr/include/qt6/QtTest/qtestcase.h:580:34: error: ambiguous overload for ‘operator==’ (operand types are ‘const KDevelop::TypePtr<KDevelop::AbstractType>’ and ‘const KDevelop::TypePtr<KDevelop::IntegralType>’)
  580 |         return compare_helper(t1 == t2, "Compared values are not the same",
      |                               ~~~^~~~~
/usr/include/qt6/QtTest/qtestcase.h:580:34: note: candidate: ‘operator==(int, int)’ (built-in)
/usr/include/qt6/QtCore/qshareddata.h:202:5: note: candidate: ‘bool operator==(const QExplicitlySharedDataPointer<KDevelop::AbstractType>&, const QExplicitlySharedDataPointer<KDevelop::AbstractType>&)’
  202 |     DECLARE_COMPARE_SET(const QExplicitlySharedDataPointer &p1, p1.d, const QExplicitlySharedDataPointer &p2, p2.d)
      |     ^~~~~~~~~~~~~~~~~~~
/usr/include/qt6/QtCore/qshareddata.h:202:5: note: candidate: ‘bool operator==(const QExplicitlySharedDataPointer<KDevelop::IntegralType>&, const QExplicitlySharedDataPointer<KDevelop::IntegralType>&)’

Compare raw pointers obtained via QExplicitlySharedDataPointer::get()
instead to resolve the ambiguity.