Refactor PersistentSymbolTable::visitFilteredDeclarations for readability
Use the IILE pattern also for the `cachedImports` and then adapt the
`filterIterator` IILE slightly to make it simpler to understand (imo).

Sadly in Qt5 we cannot yet assign a non-const iterator we get from
insertion to the const iterator we initially get from the lookup.
So this version of the patch is a bit more ugly than what I had
in mind originally. Still better than before though.