Fix PersistentSymbolTable::visitFilteredDeclarations for Qt6
Do not rely on QHash iterator stability which is not guaranteed
with Qt6 anymore. Instead, replace the QVarLengthArray with a
QVector to ensure we get a stable container we can iterate over.
We don't have to copy anything either thanks to Qt container COW.

This change was suggested by Igor Kushnir, thanks!