Fix PersistentSymbolTable::visitFilteredDeclarations for Qt6
Do not rely on QHash iterator stability which is not guaranteed
with Qt6 anymore. Instead, use a shared_ptr indirection for the
inner hash in the `declarationsCache`. That gives us a stable
location we can reference during iteration. We don't have to copy
anything either, ensuring we don't blow up the stack size
accidentally, the QVarlengthArray we have here is 1KB after all.

Note that a unique_ptr would be fine too in theory, but sadly QHash
does not support move-only value types... We don't want to use
std::unordered_map either, as that's known to be much slower than
QHash. So let's just use shared_ptr as it's easy to use and ensures
we don't leak anything accidentally.