Drop defunc support for KDEV_DISABLE_WELCOMEPAGE env var
Currently the welcome page plugin is always loaded, only controlled by
the user having endabled or disabled it in the plugin settings.

This is because 864d6dcf set "X-KDevelop-Category": "Global" also for the
welcomepage plugin, breaking the concept of using an empty category to
avoid it being loaded by the plugin controller by default.
The respective comment about the approach "Don't set a category, we want
to explicitly invoke it" added in 2c8a17d4 was lost in the desktop->JSON
metadata conversion.

With KDevPlatform now rather an internal thing of KDevelop, and also no-one
having noticed that breakage of the env control flag in years, treating the
welcomepage plugin simply as normal global plugin, without explicit custom
loading code, seems fine.