fixup QScopedPointer and QPointer porting? get() and Adapt to new in Qt 6 QPointer::operator= overload
The pointers actually are convertible without any GCC or Semantic
analysis (Clang) warnings. No idea why Aleix made these changes as part
of the aggregate commits "Port a bunch of deprecated API" and
"Few smaller porting changes".

Revert the QPointer parts because they should not be more efficient and
do not compile against Qt 5 (because QPointer::get() is missing in Qt 5;
the get()'s alias data() could in principle be used instead).

Keep the QScopedPointer part of the commits, because it is correct and
necessary.

Commit message:
RegistersManager: don't cast pointer to bool before printing

The type of m_registerController is QScopedPointer<IRegisterController>.
QScopedPointer is not convertible to its raw pointer type, but is
implicitly convertible to bool in Qt version 5 and earlier. Such a
conversion to bool happens when a QScopedPointer is printed via QDebug
in Qt 5. This printing does not compile at all in Qt 6, because
QScopedPointer::operator bool() became explicit.

The conversion to bool before printing was accidentally introduced by
the commit that changed the type of m_registerController from raw
pointer to QScopedPointer: 1821513c2de8211d4cc3736af10af7c96fba85cd.