static_cast qsizetype index to int
QStringList::indexOf() returns an int in Qt 5 but a qsizetype in Qt 6.
Cast the index to int in order to let compilers deduce std::max's
template parameter type while building against Qt 6.
QComboBox::setCurrentIndex() still takes an int in Qt 6. The number of
supported Meson backends is not going to be out of the 32-bit integer
range.