ProjectSelectionPage: port to QRegularExpressionValidator
QRegExpValidator is entirely absent from Qt 6, not even retired to the
Core5Compat module.

The only significant difference between QRegExpValidator and
QRegularExpressionValidator is that the former uses a QRegExp and the
latter - a QRegularExpression.

The default ValidProjectName pattern value has the same meaning for
QRegExp and QRegularExpression. KDevelop's built-in app templates
override ValidProjectName with only one other pattern value:
"^[a-zA-Z_][a-zA-Z0-9_-]+$". This overriding pattern also has the same
meaning for QRegExp and QRegularExpression. If external app templates
override ValidProjectName with a less straightforward pattern, let them
port to QRegularExpression syntax.