GrepOutputDelegate: port from QRegExp to QRegularExpression
The ported pattern has the same meaning for QRegExp and
QRegularExpression, except that the character class \s matches
characters with the corresponding Unicode property in QRegExp and only
ASCII characters in QRegularExpression. We do not care much about the
unlikely scenario of source code indentation using non-ASCII whitespace.
Not enough to use the pattern option
QRegularExpression::UseUnicodePropertiesOption and preserve the QRegExp
behavior.