TemplateRenderer: port from QRegExp to QRegularExpression
The ported regex is used to replace characters forbidden in a Grantlee
variable name with underscores. I have tested KTextTemplate (the KF6
replacement for Grantlee) and learned that non-ASCII characters are not
supported in KTextTemplate variable names. They are most likely not
supported in the Qt 5 version of Grantlee either. So do not use the
pattern option QRegularExpression::UseUnicodePropertiesOption.