QMakeFileVisitor: port from QRegExp to QRegularExpression
The ported regex handles the regex replacement operator ~= of the qmake
language. The qmake program, which parses qmake files and defines the
syntax, is part of qtbase. The operator ~= is parsed by the block
`if (tok == TokReplace)` of the function
QMakeEvaluator::visitProVariable() in the file
qtbase/qmake/library/qmakeevaluator.cpp. The only difference between the
Qt 5 and Qt 6 parsing is that the former uses QRegExp and the latter -
QRegularExpression. Apparently user qmake regex patterns are expected to
be compatible with QRegularExpression or users are expected to port the
patterns during the upgrade to Qt 6. Either way, let us follow suit and
use QRegularExpression in the Qt 6 version of KDevelop.