Adapt to Qt 6 QMetaType's automatic registration of stream operations
Declaring a signal, slot or Q_DECLARE_METATYPE() with
QList<GrepOutputItem> requires a default GrepOutputItem constructor in
order to compile against Qt 6. The reason for this new requirement is
the automatic registration of data/debug stream operations in QMetaType
implemented by
https://code.qt.io/cgit/qt/qtbase.git/commit?id=ed8acbeb7c6fd137486b5c8a6eec58a3b9db862c

Override QStandardItem::read() and QStandardItem::write() to detect when
a GrepOutputItem is read from or written to a stream (QStandardItem's
QDataStream operators delegate to these virtual functions). Serializing
the DocumentChangePointer m_change data member of GrepOutputItem is not
trivial, because it is an explicitly shared pointer and because the
pointed-to value DocumentChange contains an IndexedString data member.

I could not trigger a default construction, reading from or writing to a
stream of a GrepOutputItem. So do not figure out how to implement these
new member functions properly for a hypothetical future use. Instead,
print an error message and assert false in the definitions, deprecate
and warn against using in the documentations of the new member
functions.