Simplify quotation-mark handling in descriptionFromHtmlData
Don't rely on a funky string concatenation and clunky regexp for
something as simple as a character class containing a single or
double quote.

While at it, escape the fragment before using it in the regexp
and use capture groups to find matching closing quotes as suggested
by Igor Kushnir.