Port away from KF5-only KColorSchemeManager::createSchemeSelectionMenu()
This function is deprecated since KF 5.107 and gone in KF 6.

Investigating the needed changes showed also that the custom wrapper
ColorSchemeChooser around the KF classes no longer is needed, instead
the KF classes provide all the wanted config integration out-of-the-box.

The used KColorSchemeManager instance is now life-time-managed by the
MainWindowPrivate instance instead of its actionCollection.