Properly extract documentation for enumerators via QtHelp
The DUChain type system encodes their identifiers as e.g.
`Qt::GlobalColor::black` even though `Qt::GlobalColor` is not an
`enum class`. That then leads to lookup failures for the identifier
in the QtHelpEngine.

Workaround this by adding a fallback that simplifies the id.
Keep the normal logic in place in case an `enum class` is encountered
where the full id would be needed.