QtHelpDocumentation: don't match comment marker for a longer ID
The description of `typedef QString::size_type` precedes the description
of `int QString::size() const` in both Qt 5 and Qt 6 documentations.
descriptionFromCommentMarkers() looks for "<!-- $$$size", finds
"<!-- $$$size_type -->" and returns an empty description for
QString::size(), because the description of the typedef is empty.

The identifier in a comment marker can be followed by several different
non-identifier characters, such as '-', ' ', '[' and '$'. For example:
<!-- $$$QString-brief -->
<!-- $$$Iterator -->
<!-- $$$append[overload1]$$$appendconstQString& -->
<!-- $$$append$$$appendQChar -->
<!-- $$$acceptDrops-prop$$$acceptDrops$$$setAcceptDropsbool -->

Look for the next occurrence of the searched-for comment marker if the
already found marker is followed by an alphanumeric character or an
underscore, which must be part of a longer identifier.