Don't always interpret filter wildcard pattern as filepath globbing
https://code.qt.io/cgit/qt/qtbase.git/commit/?id=eb349930eee1ca8f31415dd2269a70d3dfd76257
migrated QSortFilterProxyModel::setFilterWildcard() to
QRegularExpression, and therefore * in a wildcard no longer matches /
since Qt 6.0.

Wildcards are sometimes used in non-filepath contexts, and so
https://code.qt.io/cgit/qt/qtbase.git/commit/?id=4b197c3f52b443c11e980f87aa035b81948871ce
implemented a NonPathWildcardConversion flag in Qt 6.6. The flag is not
used in QSortFilterProxyModel::setFilterWildcard() to restore the Qt 5
behavior.

Implement a helper function that works just like the Qt 5 version of
setFilterWildcard() (given Qt 6.6 or later) and use it in non-filepath
contexts. The remaining two uses of setFilterWildcard() in
CommitToolView and ScratchpadView are in filesystem contexts: file paths
and file names respectively are filtered. Neither TestView's nor
ScratchpadView's filtered items can contain slashes, so the choice
between setFilterWildcard() and setFilterNonPathWildcard() is made based
on whether they are filesystem items.