BranchManager: convert source index to proxy index
8f77f44b67c750865d0ac5e6b23ffaba52017c31 introduced the
QSortFilterProxyModel but failed to implement the necessary index
conversion. Upgrading to Qt 6 revealed this old bug, because now the
following runtime warning is printed whenever the Branch Manager dialog
is opened:
    QSortFilterProxyModel: index from wrong model passed to mapToSource

An alternative fix is to look for branchName in m_filterModel instead of
m_model. But searching in the source model guarantees consistency with
the current-branch code in BranchesListModel::refresh().

My search for more missing conversions like the one added in this commit
turned up nothing.