Really wrap the std::map pretty printers for QMap
I.e. lookup the std::map pretty printer and defer to it when
getting queried for children. Then also implement `num_children`
and try to use that to get QMap pretty printed output like we used
to get with Qt4/5. Sadly, the libstdc++ pretty printers don't yet
implement `num_children`, so we have to jump through some hoops:
I.e. when `num_children` is not available, we instead try to
extract the size from the string contents, since that is going to
be a O(1) operation. It was contemplated to sum the children iterator
to get the size, but that would be an O(N) operation and potentially
too slow for large maps.