Really wrap the std::map pretty printers for QMap
I.e. lookup the std::map pretty printer and defer to it when
getting queried for children. Then also implement `num_children`
and try to use that to get QMap pretty printed output like we used
to get with Qt4/5. Sadly, the libstdc++ pretty printers don't yet
implement num_children, so we have to jump through some hoops:

First, we try to extract the size from the string contents,
since that is going to be O(1). If that fails, we resort to a O(N)
sum over the pretty printer iterator, if possible.

If all this fails, we continue to just embed the std::map contents.