Opt-in to new gdb.ValuePrinter API and provide num_children
I.e. for those collections that are iterable, I believe it is useful
to provide the new num_children method. To let GDB know about that
we need to inherit from gdb.ValuePrinter which is relatively new,
so do that optionally. ValuePrinter reserves member names starting
with alphanumeric chars, so prefix our code with an underscore.