Cache current node in GDB QHashPrinter._iterator_qt6
Each node is computed twice during iteration over a QHash: first for the
key and then for the value. The same node would be computed repeatedly
and redundantly N+1 times for a QMultiHash key with N associated values
when Qt6 QMultiHash support is added.

Compute each node once and store the current node in a new dedicated
variable.